% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_check.R
\name{auto_check}
\alias{auto_check}
\title{Check the population by column}
\usage{
auto_check(data, na_string = c("NA", "", " ", "N/A"), check_unique = FALSE)
}
\arguments{
\item{data}{The dataframe to be checked.}

\item{na_string}{A character vector specifying the strings that should be treated as missing values. Defaults to c("NA", "", " ", "N/A").}

\item{check_unique}{Logical indicating whether to also check for uniqueness. Defaults to FALSE.}
}
\value{
A tibble containing the check results, including column names, counts of non-missing values, and percentages of non-missing values.
If \code{check_unique} is set to TRUE, the tibble also includes counts of unique values and percentages of unique values.
}
\description{
This function automatically checks for missing values and optionally uniqueness in a dataframe.
}
\examples{
auto_check(mtcars)
auto_check(mtcars, check_unique = TRUE)
auto_check(mtcars, na_string = c("NA", 999, "N/A"))

}
