% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autor-package.R
\docType{package}
\name{autor-package}
\alias{autor}
\alias{autor-package}
\title{autor: A package that showcases the possibilities of work automation with R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
