% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{Split and export data into files}
\usage{
data_split(data, folder_path, row_lim, file_names, func = readr::write_csv)
}
\arguments{
\item{data}{the data you want to export}

\item{folder_path}{the destination of the to be exported files}

\item{row_lim}{the number of row limit per exported file.}

\item{file_names}{list of names you want to give to the to be exported files. Note, they should have more unique values than the number of files exported.}

\item{func}{is the exporting method. The default is \code{func = readr::write_csv}. Note: no quote is needed and you can use any export function you have installed.}
}
\value{
A number of files containing records from data that does not exceed the \code{row_lim} per file.
}
\description{
This will take in any data frame, split and export them into multiple files.
}
\examples{
# Enable parallel processes
library(furrr)
plan(multisession, workers = 4)
data_split(mtcars,tempdir(),10,file_names = letters,func = readr::write_csv)
\dontshow{
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
