% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{is_package_installed}
\alias{is_package_installed}
\title{Helper function to check if a package is available}
\usage{
is_package_installed(pkg_name)
}
\arguments{
\item{pkg_name}{Package name to check}
}
\value{
A logical value whether a package is available.
}
\description{
Helper function to check if a package is available
}
\examples{
is_package_installed("ggplot2")
}
