% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{load_pkg}
\alias{load_pkg}
\title{Helper function to load a package if available and throw a warning when it is not}
\usage{
load_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Package to check}
}
\description{
Helper function to load a package if available and throw a warning when it is not
}
\examples{
#load_pkg(pkg_name = "tidyverse")

}
