% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{na_replace}
\alias{na_replace}
\title{Replace the strings with NAs}
\usage{
na_replace(x,na_string = c("NA",""," ","N/A"))
}
\arguments{
\item{x}{A vector.}

\item{na_string}{A character vector. It could contain any value that should be regarded as NA. The default is \code{c("NA",""," ","N/A")}}
}
\value{
A vector with NAs created by replacing the na_string values with NA.
}
\description{
This function identifies NA using the strings input.
}
\examples{
 na_replace(mtcars$carb, na_string = 2)
}
