% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{product_use}
\alias{product_use}
\title{Simulated product usage data}
\format{
\subsection{\code{product_use}}{

A data frame with 5,848 rows and 4 columns:
\describe{
\item{customers}{Customer symbols}
\item{date}{Date}
\item{usage}{Numeric value representing the product usage}
\item{region}{north, east, west, south}
...
}
}
}
\usage{
product_use
}
\description{
A simulated data for daily product usage from 8 customers across 4 regions from July 2021 to July 2023
}
\keyword{datasets}
