% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{str_after_gen}
\alias{str_after_gen}
\title{Get the string after a specific pattern (generic)}
\usage{
str_after_gen(string, pattern, position = 1, include = FALSE)
}
\arguments{
\item{string}{A character vector with 1 element.}

\item{pattern}{Pattern, string or special characters.}

\item{position}{A number. Which one to use, if multiple matches were found. The default is \code{1}. If the last one needs to be chosen but the position can range from x to y, then choose a number bigger than y and the function will choose the last one.}

\item{include}{Logical value. Whether to include the pattern, the default is \code{FALSE}.}
}
\value{
A character vector after the pattern.
}
\description{
This function extracts the string after a specific pattern.
}
\examples{
str_after_gen(c("John|09-01-22"),pattern = "\\\\|")
}
