% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_before.R
\name{str_before}
\alias{str_before}
\title{Get the string before a specific pattern}
\usage{
str_before(string,pattern,position = 1,include = FALSE)
}
\arguments{
\item{string}{A character vector. It can have multiple elements.}

\item{pattern}{Pattern, string or special characters.}

\item{position}{A number. Which one to use, if multiple matches were found. The default is \code{position = 1}. If the last one needs to be chosen but the position can range from x to y, then choose a number bigger than y and the function will choose the last one.}

\item{include}{logical value. Whether to include the pattern, the default is \code{include = FALSE}.}
}
\value{
A character vector after the pattern.
}
\description{
This function extracts the string before a specific pattern.
}
\examples{
str_before(c("John|09-01-22","Jane|09-02-22"),pattern = "\\\\|")
}
