% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{ticket_sales}
\alias{ticket_sales}
\title{Simulated ticket sales data}
\format{
\subsection{\code{ticket_sales}}{

A data frame with 50,900 rows and 4 columns:
\describe{
\item{date}{Date}
\item{theater}{Different theaters}
\item{state}{States that the theater is located}
\item{sales}{Daily sales}
...
}
}
}
\usage{
ticket_sales
}
\description{
A simulated data for daily theater ticket sales from 100 theaters across 50 states from April 2022 to Aug 2023.
}
\keyword{datasets}
