% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{update_ticket_sales}
\alias{update_ticket_sales}
\title{Helper function to update the date column from the ticket_sales dataset to reflect recent days}
\usage{
update_ticket_sales(ticket_sales)
}
\arguments{
\item{ticket_sales}{The ticket_sales dataset}
}
\description{
Helper function to update the date column from the ticket_sales dataset to reflect recent days
}
\examples{
#ticket_sales_new <- update_ticket_sales(ticket_sales)
}
