% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{value_count}
\alias{value_count}
\title{Helper function to create a summary table for the \code{auto_check()}}
\usage{
value_count(x, value = c("unique","missing","non_missing"))
}
\arguments{
\item{x}{A vector.}

\item{value}{A vector of summary options. It includes unique count, missing count and non-missing count values. The default is \code{c("unique","missing","non_missing")}}
}
\value{
A data frame with summary information about vector x.
}
\description{
This function creates the unique count, missing count and non-missing count values in a data frame for a vector.
}
\examples{
value_count(mtcars$mpg)
value_count(mtcars$mpg,value = "non_missing")
}
