% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merger}
\alias{merger}
\title{Merge and load multiple files in your R session}
\usage{
merger(folder_path, pattern = "", func = readr::read_delim, show_name = FALSE, para = FALSE, ...)
}
\arguments{
\item{folder_path}{the path to the folder that holds the target files}

\item{pattern}{the pattern that helps identifies the target files. This will take all the regex or keywords or unique characters. The default is \code{pattern = ""}}

\item{func}{the import method of your choice. The default is \code{func = readr::read_delim}. Note: no quote is needed and you can use any import function you have installed, for example, \code{func = readr::read_csv} if you have installed \code{readr} package.}

\item{show_name}{logical value. Whether we have to have the merged file to append the filename. The default is \code{show_name = FALSE}}

\item{para}{logical value. Whether we need to utilize parallel processes with the \code{furrr} package to import the files faster. The default is \code{para = FALSE}}

\item{...}{Any argument you want to pass to the import function. For example, you can pass \code{n_max = 100} to limit the import to be 100 lines per files if you are using import functions from \code{readr}.}
}
\value{
a merged data frame for all targeted files.
}
\description{
This will identify all files in a folder that meet the criteria set and import them into you session.
}
\examples{
folder_path <- tempdir()
data_split(mtcars,folder_path,10,file_names = letters)
merger(folder_path, pattern = "csv", show_name = TRUE)
# Enable parallel processes
library(furrr)
plan(multisession, workers = 4)
merger(folder_path, pattern = "csv", show_name = TRUE, para = TRUE)
\dontshow{
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
